+++ packages/react/Link/src/components/Skeleton/skeleton.types.ts
@@ -1,0 +1,5 @@
+import { SkeletonProps as SkeletonBaseProps } from "@nimbus-ds/skeleton";
+
+export type SkeletonProps = Partial<Pick<SkeletonBaseProps, "width">> & {
+  "data-testid"?: string;
+};