+++ packages/styles/webpack.config.js
@@ -28,6 +28,7 @@ module.exports = {
     "./packages/list/index": "./src/packages/list/index",
     "./packages/radio/index": "./src/packages/radio/index",
     "./packages/toggle/index": "./src/packages/toggle/index",
+    "./packages/link/index": "./src/packages/link/index",
     "./packages/tooltip/index": "./src/packages/tooltip/index",
   },
   output: {